refactor: narrow report document API
