+++ apps/server/tests/adapters/pdf/test_i18n_separation.py
@@ -17,7 +17,8 @@ import pytest
 from vibesensor.adapters.analysis_summary import summarize_run_data
 from vibesensor.report_i18n import is_i18n_ref, tr
 from vibesensor.report_i18n import resolve_i18n as resolve_i18n_impl
-from vibesensor.use_cases.history.report_document import build_report_document, prepare_report_input
+from vibesensor.shared.boundaries.reporting import prepare_report_input
+from vibesensor.use_cases.history.report_document import build_report_document
 
 _TRANSLATED_MARKERS = [
     # Dutch markers
