+++ apps/server/tests/adapters/pdf/test_report_content_coverage.py
@@ -23,6 +23,7 @@ from vibesensor.adapters.analysis_summary import summarize_log
 from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
 from vibesensor.domain import Finding
 from vibesensor.shared.boundaries.finding import finding_from_payload
+from vibesensor.shared.boundaries.reporting import prepare_report_input
 from vibesensor.shared.boundaries.reporting.document import (
     AppendixAData,
     AppendixCData,
@@ -37,7 +38,7 @@ from vibesensor.shared.boundaries.reporting.document import (
 )
 from vibesensor.shared.constants.units import KMH_TO_MPS
 from vibesensor.use_cases.diagnostics.top_cause_selection import select_top_causes
-from vibesensor.use_cases.history.report_document import build_report_document, prepare_report_input
+from vibesensor.use_cases.history.report_document import build_report_document
 
 _I18N_JSON = SERVER_ROOT / "data" / "report_i18n.json"
