+++ apps/server/tests/adapters/pdf/test_report_metrics_consistency.py
@@ -35,8 +35,9 @@ from test_support.sample_scenarios import (
 )
 
 from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
+from vibesensor.shared.boundaries.reporting import prepare_report_input
 from vibesensor.shared.boundaries.reporting.document import ReportDocument
-from vibesensor.use_cases.history.report_document import build_report_document, prepare_report_input
+from vibesensor.use_cases.history.report_document import build_report_document
 
 # ---------------------------------------------------------------------------
 # Type alias
