+++ apps/server/tests/adapters/pdf/test_report_pdf_diagram_hotspots.py
@@ -472,10 +472,8 @@ def test_build_report_pdf_hotspot_panel_explains_intensity_and_certainty() -> No
     from test_support.report_helpers import minimal_summary
 
     from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
-    from vibesensor.use_cases.history.report_document import (
-        build_report_document,
-        prepare_report_input,
-    )
+    from vibesensor.shared.boundaries.reporting import prepare_report_input
+    from vibesensor.use_cases.history.report_document import build_report_document
 
     summary = minimal_summary(
         lang="en",
