+++ apps/server/tests/adapters/pdf/test_report_strength_label_peak_amp.py
@@ -5,7 +5,8 @@ from __future__ import annotations
 from test_support.report_helpers import minimal_summary
 
 from vibesensor.adapters.pdf.pdf_drawing import _strength_with_peak
-from vibesensor.use_cases.history.report_document import build_report_document, prepare_report_input
+from vibesensor.shared.boundaries.reporting import prepare_report_input
+from vibesensor.use_cases.history.report_document import build_report_document
 from vibesensor.use_cases.history.report_document.presentation import strength_text
 
 # ---------------------------------------------------------------------------
