+++ apps/server/tests/adapters/pdf/test_summary_view.py
@@ -14,13 +14,13 @@ from vibesensor.domain import (
     TestPlan as DomainTestPlan,
 )
 from vibesensor.shared.boundaries.analysis_summary_projection import project_analysis_summary
+from vibesensor.shared.boundaries.reporting import prepare_report_input
 from vibesensor.shared.boundaries.reporting.projection import resolve_report_origin
 from vibesensor.shared.boundaries.test_plan_projection import step_payloads_from_plan
 from vibesensor.shared.boundaries.test_run_reconstruction import (
     test_run_from_summary as _test_run_from_summary,
 )
 from vibesensor.shared.boundaries.vibration_origin import origin_payload_from_finding
-from vibesensor.use_cases.history.report_document import prepare_report_input
 
 
 def _minimal_summary(**overrides: object) -> dict[str, object]:
