+++ apps/server/tests/integration/test_report_pipeline.py
@@ -41,7 +41,8 @@ from test_support import (
 )
 
 from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
-from vibesensor.use_cases.history.report_document import build_report_document, prepare_report_input
+from vibesensor.shared.boundaries.reporting import prepare_report_input
+from vibesensor.use_cases.history.report_document import build_report_document
 
 # ---------------------------------------------------------------------------
 # Helpers
