+++ apps/server/tests/use_cases/diagnostics/test_analysis_pipeline_integration_regressions.py
@@ -17,12 +17,13 @@ from vibesensor.adapters.persistence.history_db import HistoryDB
 from vibesensor.shared.boundaries.persisted_analysis_codec import (
     persisted_analysis_to_json_object,
 )
+from vibesensor.shared.boundaries.reporting import prepare_report_input
 from vibesensor.shared.boundaries.run_log import normalize_sample_record
 from vibesensor.shared.boundaries.run_metadata_codec import run_metadata_from_mapping
 from vibesensor.shared.boundaries.sensor_frame_encoder import sensor_frame_to_json_object
 from vibesensor.shared.sampling import bounded_sample
 from vibesensor.shared.types.run_schema import RunMetadata
-from vibesensor.use_cases.history.report_document import build_report_document, prepare_report_input
+from vibesensor.use_cases.history.report_document import build_report_document
 from vibesensor.use_cases.run import RunRecorder, RunRecorderConfig
 
 # ---------------------------------------------------------------------------
