+++ apps/server/tests/use_cases/diagnostics/test_diagnosis_robustness_report_pdf.py
@@ -13,7 +13,8 @@ from test_support import (
 )
 
 from vibesensor.adapters.analysis_summary import summarize_run_data
-from vibesensor.use_cases.history.report_document import build_report_document, prepare_report_input
+from vibesensor.shared.boundaries.reporting import prepare_report_input
+from vibesensor.use_cases.history.report_document import build_report_document
 
 
 class TestPdfContentForDiagnosedScenario:
