+++ apps/server/tests/use_cases/history/test_report_facts.py
@@ -4,7 +4,7 @@ import pytest
 from test_support.findings import make_finding_payload
 from test_support.report_helpers import minimal_summary
 
-from vibesensor.shared.boundaries.reporting import prepare_report_facts
+from vibesensor.shared.boundaries.reporting import prepare_report_facts, prepare_report_input
 from vibesensor.shared.boundaries.reporting.summary import report_summary_from_mapping
 from vibesensor.shared.boundaries.test_run_reconstruction import (
     test_run_from_summary as build_test_run_from_summary,
@@ -12,7 +12,6 @@ from vibesensor.shared.boundaries.test_run_reconstruction import (
 from vibesensor.shared.run_context_warning import RunContextWarning
 from vibesensor.use_cases.history.report_document import (
     compose_report_document_context,
-    prepare_report_input,
 )
